refactor(gnrwebstruct): nest _widgets package inside gnrwebstruct
Moves `gnr/web/widgets/` to `gnr/web/gnrwebstruct/_widgets/` so the
declarative widget catalog lives next to the dispatcher that consumes
it. The leading underscore signals package-private status, consistent
with the existing `_helpers.py` in the same package.

- 5 files renamed (git mv): `__init__.py`, `html.py`, `dijit.py`,
  `dojox.py`, `genro.py`
- Internal imports updated to use the new dotted path
- `dojo11.py`: `from gnr.web.widgets` -> `from gnr.web.gnrwebstruct._widgets`
- `tests/web/widgets_test.py`: same import update + docstring
- `pyproject.toml`: replace `gnr.web.widgets` with
  `gnr.web.gnrwebstruct._widgets` under `[tool.setuptools].packages`
- No callers outside this package reference `gnr.web.widgets`
  (verified with grep across gnrpy and projects).

Tests
-----
- pytest tests/web/widgets_test.py + gnrwebstruct_test.py -> 42/42 pass
- pytest tests/                                            -> 1767 pass / 22 pre-existing fail
- flake8 gnr/web/gnrwebstruct/ tests/web/                   -> clean

Refs #901
