+++ gnrpy/gnr/web/gnrwebstruct/_widgets/dojox.py
@@ -4,7 +4,7 @@ Formal documentation of the dojoxNS widgets available on GnrDomSrc.
 These methods are generated at runtime via __getattr__ -> child(tag, **kwargs).
 """
 from __future__ import annotations
-from gnr.web.widgets import WidgetMixinBase, element
+from gnr.web.gnrwebstruct._widgets import WidgetMixinBase, element
 
 
 class DojoxWidgets(WidgetMixinBase):
