+++ gnrpy/pyproject.toml
@@ -158,8 +158,8 @@ packages = [
 	    'gnr.web.gnrwebpage_proxy.apphandler',
     	    'gnr.web.gnrwebpage_proxy.frontend',
 	    'gnr.web.gnrwebstruct',
+	    'gnr.web.gnrwebstruct._widgets',
 	    'gnr.web.gnrwsgisite_proxy',
-	    'gnr.web.widgets',
 	    'gnr.xtnd',
 	    'gnr.xtnd.cli',
    	    'gnr.lib',
